+++ src/jobs.js
@@ -1,0 +1,6 @@
+import {cozyFetchJSON} from './fetch'
+
+export function count (cozy, workerType) {
+  return cozyFetchJSON(cozy, 'GET', `/jobs/queue/${workerType}`)
+    .then(data => data.attributes.count)
+}